+++ etix-admin/src/components/ServiceMan/Services.jsx
@@ -404,9 +404,16 @@ const Services = () =>{
     };
 
     const handleDelete = (ids) => {
-        ids.map((id) => {
-            dispatch(deleteService(id));
-        })
+        var dlt = window.confirm("All data related with this service will be deleted. Are you sure to delete instead of setting the status to inactive? ")
+        if(dlt){
+            ids.map((id) => {
+                dispatch(deleteService(id));
+            })
+        }
+        else{
+            return;
+        }
+        
     }
     
     return (
