+++ etix-admin/src/components/User/UserDetail.jsx
@@ -271,7 +271,14 @@ const UserDetail = ({props}) => {
             return;
         }
         else{
-            dispatch(deleteUsers(id));
+            let dlt = window.confirm("All data related with this user will be deleted. Are you sure to delete instead of setting the status to inactive?")
+            if(dlt){
+                dispatch(deleteUsers(id));
+            }
+            else {
+                return;
+            }
+            
         }
 
         alert("Sucessfully Deleted");
